new members & traders moved to analytics, wallet owned tokes endpoint added
