+++ src/api/routes/walletRoutes.js
@@ -9,4 +9,5 @@ walletRouter
     .get("/test", walletController.test)
     .get("/test-smart-account/:privateKey", walletController.test_smart_account_though_private_key)
     .get("/test-encryption-decryption/:privateKey", walletController.test_encryption_decryption_by_private_key)
-    .get("/smart-wallet-details/:walletId", walletController.get_smart_wallet_details);
+    .get("/smart-wallet-details/:walletId", walletController.get_smart_wallet_details)
+    .get("/tokens/:walletId", walletController.get_wallet_tokens);
